Shaper use full bandwidth units of measure
In 2.2.* the shaper bandwidth type dropdowns show "Bit/s", "Kbit/s"... but in 2.3-BETA they are just showing "b", "Kb"...
That is fixed in the various array() definitions here.
I also added GetBwscaleText() function that will return the full text "Bit/s" etc associated with the current bandwidth type. Then call that in a couple of places that were displaying just the "raw" bandwidthtype on the screen. e.g. firewall_shaper_queues.php was displaying just "Kb" etc and now it will display "Kbit/s". That "issue" was in 2.2.* also - so that is an "enhancement".