print_info_box success
In most situations we want the print_info_box() of $savemsg to be
'success'. If there was some problem saving, then it can be 'warning'.
These are a few places I have found so far that need this sort of fixing
up - more to come but sleep calls just now!
I made the change to index.php also to be consistent with the others,
even though it does not actually change anything "for real" because
'warning' is the default anyway. It will make it easy for people in
future to see that is the case, specially if they are adding other
$savemsg things that need to be 'success'.