services_dnsmasq put info in infoblock
There are 3 sets of info. But when I put each in an infoblock then:
1) Only the first "i" icon works to open and close, and it opens and closes all 3 blocks of text.
2) The first 2 text boxes still have an "x" and they can be dismissed.

Do we want to enhance the supporting code for infoblock so that it can work with multiple infoblocks on the same page?

Or do we have a policy that there is only to be 1 infoblock on a page, and other 'info' text just has to not be in an infoblock?