IPsec phase 2 display existing keylen
When you edit an IPsec phase2, if there are any particular key lengths already selected in the config for each of the encryption algorithms they are not displayed - they just all show as Auto. That then "accidentally" changes previous settings when you save.

Found as part of searching for dodgy-loking uses of "==".