Consistent use of System Update
In some places it said "System update" and in others "System Update" - be consistent.
The confirm button called it "Upgrade". Make that consistent with the existing use of the term "Update" for system updates.

Note: Packages are still said to be "upgraded" rather than "updated" - there are places where headings, confirm prompts etc will say "Upgrade package".
Is that intended? That the terminology for packages is to "upgrade" them and for the base system it is an "update"?