diag_confbak make a single gettext
Sentences that are constructed in pieces like this are a potential problem for translation, if the target language needs to use a different grammatical word order. e.g. in Nepali we would say datetime-dekhi datetime-samma (datetime-from datetime-to).
Organize it into a single gettext() that has the whole sentence in one go with numbering of the parameters so they can also be potentially re-ordered as needed in a translation.
I am having a look at a few places where the gettext() translation is missing, and noticed this kind of thing also. When I see this sort of sentence construction using sentence fragments built up with multiple gettext() calls I will fix them up like this and make PRs.