Internationalization of diag_defaults
A couple of typical fixes here:
1) Actually wrap the "Are you sure...?" question in gettext()
2) Use a single gettext() call combined with sprintf() and %s etc to construct a sentence in a single go.
3) Use %s rather than just appending a variable at the end of a sentence of text. That way there is no assumption that the variable text must go at the end in every language.

Once I have got the feel of the variations I come across, I will make myself a branch and change a whole bunch of files. But for now I have submitted a few things individually so it is easy for people to comment and give any helpful advice on how to do it better, before I go editing lots of files.

IMHO it will be good to cleanup and optimize the internationalized code, then life should be easier for future translators.