Fix easyrule fbegin.inc echoed text
If you manually navigate directly to easyrule.php you get a bit of rubbish text displayed before the actual message about "This is the Easy Rule status page...". That is due to having too many closing PHP tags. Also, fbegin.inc does not need to be included anyway.

I also changed the message to be in a print_info_box() warning and sorted the internationalization a bit more.

There could be a bit more formatting done here after this - the page has a help icon that looks a bit lonely without the other usual panel-heading panel-title stuff.