Fix #5778 do not save changes if interface description matches an alias name
See https://redmine.pfsense.org/issues/5778 for details of how to reproduce the problem.

Note that similar code to make the "Sorry, an alias with the name XXX already exists" message is also at the top of interfaces.inc - it compares the current interface descr from the config with the currently existing alias names. That check would help warn the user if someone managed to add an alias name that matched the interface name. I guess it was there from some time in the past when the alias edit code did not cross-validate the alias name with the interface descriptions. I have left that check there - it does no harm to have it "just in case".

The new code that I added checks the proposed interface description in $_POST against the existing alias names and will give an input_error if there is a match.