Remove print_info_box_np
Put the real code into print_info_box, adding just a $btntext parameter
to the existing 3 params of print_info_box.
Change the remaining print_info_box_np() calls to use print_info_box()

Note: The previous code also had a "bug" - the "apply" button said
"apply" literally on the UI, when it should have said "Apply changes".
That is fixed so that now the "Apply changes" or other specified button
text is actually displayed to the user. (In the old code it was
literally displaying $name when it should have been $value)