Use template string for JS countdown in diag_reboot
Tested on Windows10 with:
Chrome 47.0.2526.111 m
Firefox 38.5.2
Edge (current with all Windows 10 updates)

This will allow translators to translate a whole string:
'Page will automatically reload in ${time} seconds.'
into some different language word order, retaining the text of the
${time} reference, like:
'Page ${time} seconds in automatically will reload'
and that should get JS substitution of the time var as it counts down in
real-time.
This provides similar functionality to PHP
sprintf('Page will automatically reload in %s seconds.', $time);

Please test, and if this seems like the way to go for this sort of
thing, then we can do similar in other places that have JS vars that are
variable at run-time and need to be substituted into sentences at
run-time.