Standardize print_apply_box usage
1) For the translators, it should be better to do it consistently. Then
they only need to have 1 translation of "You must apply the changes in
order for them to take effect" and can translate each of the other
custom sentences individually.
2) The full-stop should always go at the end of the sentence, inside the
gettext(). There are some languages where the full-stop is not the same
as European. e.g. In Nepal we use the vertical bar for end of sentence.