Fixup old diag_logs references
Note: The shortcuts for upnp were not showing. This was because
shortcuts.inc only includes *.inc files from /usr/local/www/shortcuts
So I have renamed pkg_upnp.php to pkg_upnp.inc
I notice that some time in the past it used to be pkg_upnp.inc because
that was in the obsoleted files list. So not sure what that was all
about.