Internationalize status_graph_cpu
This one is a bit odd, so I have made a separate pull request. There seems to be no place in the menus that comes to this page - I had to manually enter the URL to get to it to test. I checked on 2.2.6 also, and there is no menu entry for it.

But the code does work.

Should this be available in the menus somewhere?