status_ntpd whitespace
Various lines were not on tab stops, some PHP end tags were on the end of lines and others on new lines... I found it a bit hard to read when tracking through a previous missing PHP start tag. Maybe this is better? or maybe I have gone too crazy re-formatting it?