Status IPsec info box
As well as internationalization of this info box, I have:
1) Made all of them the same text - it seemed odd that there were 2
different versions of text for the same thing.
2) Made it an info-level infoblock. It seemed odd to me that this was
showing as a warning - it is really just a helpful info with link.
3) Made it open on page load if IPsec is not enabled, otherwise if IPsec
is already enable it loads closed and the user can open and read it if
they care.

What do you think about this combination of things in the UI?

I suspect there are a few other places where there is a convenient link
'You can configure xxx here', but it shows as a warning.
If there is agreement about the infoblock approach I have suggested
here, then I can do similar in other places.