Limit interface MTU to 1500 if no explicit MTU configured
This fixes a regression introduced by 5815d13659250c8319424268a1c8836928cd8cbf . Following that commit, VLANs on the same parent interface as a PPPoE parent interface finish up with an MTU of (PPPoE MTU + 8) if no MTU is configured explicitly. A LAN MTU of 1508 without explicit configuration was rather unexpected!