Fix #3209 editing unassigned VLAN tag can change an assigned interface
To reproduce:
a) Create VLAN 10 on some real device (e.g. em0)
b) Create VLAN 11 similarly.
c) Assign VLAN 11 (the last one in the vlan array in the config) to OPT1.
d) Edit VLAN 10 tag to be 12.
The OPT1 assignment becomes to em0 tag 12.

This is because of the wrong reference to $vlan['vlanif'] that is fixed here. At this point $vlan is a leftover from a previous loop that went through all the VLANs in the config - so it happens to be the last one from the config.

Hopefully this fixes all the various ways people have reported this in redmine #3209.