Bring all calls of print_info_box to same standard
The call itself to print_info_box already echoes the content. There is no need of additionally using the short-open-echo tag on those calls to echo return value. The previous implementation shouldn't yell any visible issues as return is 'NULL' (undefined) which casts to an empty string when printing. But, just for the sake of conformity, this changes are advisable in my opinion.