Allow user to dismiss diag_edit messages
In particular, I like to dismiss the "File saved successfully." message. Then when I save again I can see that the message re-appears.
Also, when you put a wrong filename and you get a "nosuch file" or similar message, then use "Browse", that does not refresh the page. So you end up successfully browsed to somewhere but the danger message is still sitting there. So it is nice to be able to choose to dismiss a danger message also.