Collapsible Panels JavaScript Error - Fixes 5926
Not the most elegant solution but should be functional.
Would prefer each widget to have an id attribute with something unique to match that identifies it as a widget panel.  Or test if 'updateWidgets' is defined as a function since it should only be so on the index page (dashboard).  But that seems sort of a hackish means.

Changing the widgets available panel href to "#widgets-available_panel-body" (note the added 's' to widget) may be the next best thing so that only the one if match test would be needed.

Is a call to updateWidgets even need here?  It's not readily obvious to me what it's accomplishing here when collapsing/expanding a widget panel.  My guess is that the intention is to save the panel collapsible state.  But I'm not experiencing that.