Think you are saying to put the second info block on a new line.  Though I could live with that.  I'm not fond of the extra dedicated line.  Consuming vertical space when there is plenty of unused horizontal space available in which it flows well.
Seems like maybe the the browser is not honoring something about the info block css, such as a width, or perhaps something that hasn't been specified in the css, and causing it to line wrap.

Anyway here's a commit with the comments info block on a separate line.  I'll leave it up to you to use or not.