Improve test for firewall rule advanced setting
You could make some advanced settings, save the firewall rule, then when editing again the advanced section would not be automatically open. That makes it hard for users to realise that a rule has some advanced setting(s).
This change enhances function is_aoadv_used() so it checks all the things that are an advanced setting.