Fix $POST checks so they do not depend on string Save
This prevents problems when Save is translated to other languages.
Note: This is the only remaining place I could find with this issue.
There are other pieces of code that do:
```
if ($_POST['save'] == gettext("Save"))
```
which also works, but looks a bit ugly.
I left those alone, since they do work OK in any language.