HTML Compliance - Firewall / Schedules / Edit - Readonly Attribute
Error: Bad value readonly for attribute type on element input.

Readonly is not a valid type attribute value.
If the intention here is to prevent the fields from being edited. Use the readonly attribute.
Since editing was not actually being prevented by this I did not add the readonly attribute so as to not change existing behavior.