Add field to specify source-hash key
The source-hash pool option uses a hash of the source address to
determine the translation address. This hashing algorithm is also fed a
key, which unless specified defaults to a random value. This random
value is then generated each time pf is reloaded.

This commit adds the ability to specify the key in order to provide
consistent hashing, even when pf is reloaded.