servise_dhcp include btnadvopts in section
Before this change, $section is added to $form before btnadvopts is put into $section. The previous code looked to me like btnadvopts would end up in a $section object that got overwritten by the new Form_Section at line 1165.
This change makes sure that the button is added to the section, and then the section is added to the form.

But actually the old code is working OK. (And the new code also works just the same)

So am I missing something?