Redmine #5994 Enhance test for PPP advanced defaults
If pppoe_monthly, pppoe_weekly, pppeo_daily or pppoe_hourly was already set, then when editing the PPP interface the advanced fields were not automagically displayed. They should be, because the user needs to see these settings if they are there.

Add these extra checks to the test makes the Advanced fields automagically open correctly in this case.