Force the existence of bandwidth on root queue.
If the bandwidth is not set the speed of root queue is set by media status.

In this case a ruleset which contains queues with speeds higher than 100Mb/s will fail to be loaded when the root interface is connected to a fast ethernet port (100Mb/s) - but works fine on a gigabit ethernet port.

With the root queue bandwidth, sum the bandwidth of all subqueues to make sure we do not produce a broken ruleset.

Ticket #5721