Alias Tables RAM Disk Store
If ramdisk is enabled keep a copy of the alias tables to restore at boot time.  Otherwise unpredictable behavior may occur due to some aliases not being available when the firewall rules load.

Because alias tables are typically somewhat static, the following strategies are employed to keep write cycles to a minimum for SSD and flash drive type devices friendliness.
1) Back up during reboot/shutdown only if a backup copy of the alias table does not already exists.  This is typically during the reboot when ramdisk is first enabled.
2) Update the backup copy only when the alias table is updated with a new download, typically 1 or more days, as configured in the firewall alias.