Fix #6173 Use correct list of special nets
is_specialnet() was only using a smaller list of "special nets", so when you have a VIP as the selected special destination it was not matching it up on edit, and thus "forgetting" the special VIP selection.
build_dsttype_list() already makes a good list of special destination types. That list includes all the "special" VIPs.