Fix #6251 Temperature on System Information Widget
It just seems that some infrastructure for a temperature bar had been put in place, but the code never finished to actually do it. But I am a bit surprised that it has taken so long for someone to report this.
The temperature is in deg Celsius, so a 0 to 100 "progress bar" works nicely for the typical reasonable temperature range. I have not tried to make it have any sort of good/warning/really hot color indications... - that is for the Temperature Sensors widget.
I also changed to using the &deg way to get a degree symbol, rather than the various special character codes that were there. Hopefully &deg should be a fully portable way to specify this.
(cherry picked from commit eb442fec8bd079760a53372bd9eaed7c502f2e98)