Fix #6395
Port tables do not seem to cope with having comments in them when used in pf.
This change puts back the editing (sed) for ports tables that was in RELENG_2_2 and moves the different editing (leaving comments) that is done in RELENG_2_3 so that it is only applied for ordinary URL table aliases.
Hopefully it will fix the reported issue. Needs to be checked/tested.