Feature #6388 custom profile per user
This allows the admin to create/edit a user and check the new
customprofile box. The various GUI preferences are then displayed. The
admin can choose the various GUI preferences for that user - CS file
(theme), top navigation, dashboard columns, and a host of checkboxes
that turn on/off the other preferences that are available for the GUI.
That user then gets the specified GUI preferences, rather than the
system settings.
When that user changes their dashboard widgets and widget layout, that
is saved for that individual user also.
Note: This feature is only for local users. If a user is authenticated
remotely then this code provides no way to specify custom options for
the user.