Add page to manage user profile
Add a "User Profile" page to the System menu.
The menu entry is only shown if the user has "customprofile" set.
The page allows the user to edit their own custom profile settings.
Note: selected widgets and their custom layout in the dashboard columns
is already done automagically for users with "customprofile".