Move custom widget settings into user array
Rather than have a separate 'profiles' array that just had the 'widgets'
layout/settings in it by username, put the 'widgets' layout/settings
into the 'system' 'user' entry for the relevant user. That seems cleaner
to keep all user settings there in the one place. It also means there is
no need for special code to delete or rename stuff in a 'profiles'
section - everything about a user is stored in the 'system' 'user'
entry.