Fix #6468 Do not allow edit of day and times
in rows of time ranges for a schedule.
The code was always intended that the user uses the calendar pad and start hour/minute stop hour/minute drop-down fields to enter days and time range. If an existing day-time-range is wrong, then the workflow is to delete the row and then enter the correct day-time-range using the calendar pad and start hour/minute stop hour/minute drop-down fields.

I left the Description field as editable - there is not harm here in letting the user fixup text in any of the dsecriptions for each row.