Check IP Services
Configurable check ip services.
Used by Dynamic DNS services to check ip addresses.
Configure as many as check ip services as desired.  The first (highest in the list) that is enabled will be used as the default.
A future possibility could be to add an option to the Dynamic DNS entires to use a specific check ip service.
Addresses bug #6493