Check IP Services
Configurable check IP services.
Used by Dynamic DNS services to check IP addresses.
Configure as many check IP services as desired.  The first (highest in the list) that is enabled will be used as the default.
A future possibility could be to add an option to the Dynamic DNS entires to use a specific check IP service.
Addresses bug #6493 by allowing additional check IP service to be added and selected as the default.