Fix #6050 DHCP - provide Network Booting display/hide advanced button
This keeps the section header, but instead of a +/- on the right of the section header bar, it puts a Display/Hide Advanced button, like other parts of this GUI page.
The section header looks difficult to remove, because the "Additional BOOTP/DHCP Options" above has its own section, which has to end. So I am leaving the "Network Booting" section header.