dhcp6 init before RA
Some ISP BNG's require a dhcp6 init be sent before RA. If this is not
sent then no RA is acknowledged. The changes to interfaces.php add an
option to WAN DHCP6  settings and a flag to the config file -
'dhcp6withoutra'. This flag is then used in interfaces.inc to decide
whether to let rtsold launch dhcp6c as is the defualt, or to launch
dhcp6c without rtsold. There are also additions to standardise the
killing of the dhcp6c process  - bringing it into line with the dhcp4
client method.