DHCP6 init before RA
Some ISP BNG's  require that a dhcp6c init is sent before RA. The
additions to interfaces.php adds a new option flag 'dhcp6withoutra',
this flag is set or unset in the WAN DHCP6 section 'Do not wait for a
RA'.

The interfaces.inc file handles this flag and either launches dhcp6c on
interface_dhcpv6_configure or if the flag is not set will rtsold will
launch dhcp6c on receipt of RA.

An extra addition is the new function kill_dhcp6client_process($wanif).
This is added to try and create a more robust termination of and dhcp6c
clients that may be running.