Fix "http://pf.sense/UNKNOWN" links in Pkg Manager
I noticed recently that many packages do not have the 'www' field in the database filled, or have it set to UNKNOWN  The way the table is built, this causes bad links to be generated pointing to e.g. http://pf.sense/UNKNOWN.  This patch fixes it, causing only packages with actual links to get the <a> tag.