Ticket #4209: Send DHCP Release packet
When you hit 'Release' on the status_interfaces.php the dhclient did not
send the proper release packet to the DHCP server. This is sometimes
required by some ISP's and the freebsd dhclient doesn't do it, however I
found a quick fix. A copy of ISC's dhclient is sitting on pfSense 2.3 so
we will utilize it for this function.