Status > Interfaces [Update 1]
A few minor touchups on the wording along with showing the configuration
type after the IP address so you know where the IP was obtained from,
similar to DD-WRT/Tomato, etc. It also shows the CIDR notation after the
subnet masks since pfSense refers to the CIDR notation everywhere else.
It also shows the name of the NIC cuz I thought it was interesting to
show there.
https://forum.pfsense.org/index.php?topic=113088.msg629905#msg629905