Service Running Status Indicator Icon (fa-play)
Perhaps static play icon would be better perceived as the contrast to static stop icon for for services status indicators.

Regardless of which icons are chosen, this centralized services status approach provides consistency, and the status icons, thumbs up/down, and title text can be easily changed in one place.