DHCP6C without RA - Corrections and additions
Correction to the RTSOLD script creation as the existing code was allowing dhcp6c to be launched on receipt of RA when it was in dhcpwithoutra mode. This results in multiple dhcp6c clients,

kill_dhcp6client_process() now uses mwexec("kill -9 $pid"); rather than posix_kill() as the later is not reliable, the former kills the process every time.

In interfaces_bring_down(*) the pid find and posix_kill is replaced with the call to kill_dhcp6client_process()

Style is now - hopefiully  - correct!